backend/requirements.txt
Django==5.0.1
django-ninja==1.1.0
psycopg[binary]>=3.2.0
python-decouple==3.8
dj-database-url==2.1.0
django-cors-headers==4.3.1
Pillow>=10.0.0
gunicorn==21.2.0
whitenoise==6.6.0
django-extensions==3.2.3
PyJWT==2.8.0
cryptography==42.0.2
celery==5.3.4
redis==5.0.1
django-storages[google]==1.14.2
requests==2.31.0
python-multipart==0.0.6
email-validator==2.1.0
pytest==8.0.0
pytest-django==4.8.0
coverage==7.4.1
black==24.1.1
flake8==7.0.0
isort==5.13.2 